I've moved files around to make a simpler file structure.
