+++ client/src/components/Bubbles Components/BubblePage.js
