+++ client/src/components/Navigation/PrivateRoute.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Route, Redirect } from 'react-router-dom'
-import { getToken } from '../utils/api';
+import { getToken } from '../../utils/api';
 
 const PrivateRoute = (props) => {
     const {
